Adding the initial versions
Ball_Tracker.py takes a live image from a webcam andt then converts it to
HSL space to extract only the orange range to find a ball. The countour
is found to then track the center of the ball, or largest orange object.
This data is recorded with an elapsed time, then written to a csv.

Grapher,py takes the csv file that Ball_Tracker.py creates and graphs
the points on a plotter. A slider is on the bottom to adjust the time at
which the ball data is being tracked, to show the path of the ball.